Add border rendering
Currently there's a glaring issue with the way
boxes are rendered. There is no capability to
render rounded corners as that would be shader
code.

The library as it stands has no rendering
capability.

The current implementation of border rendering
is just creating 4 rects with a specified width
at the edges of the bordered box. This too is
incompatible with rounded corners.