Add border rendering
Currently there's a glaring issue with the way
boxes are rendered. There is no capability to
render rounded corners as that would be shader
code.

The library as it stands has no rendering
capability.

The current implementation of border rendering
is just creating 4 rects with a specified width
at the edges of the bordered box. This too is
incompatible with rounded corners.

Amendment:

This could actually be fine for rounded corners
assuming they are handled in the shader code.

If the border rects are supplied as such, then
rounding could be done on the top and bottom
border rects, denoted by bt and bb, respectively.

Then, bl and br do not require special handling.

|   bt   |
----------
|bl|  |br|
----------
|   bb   |